+++ html/index.html
@@ -6,7 +6,7 @@
     @import "/css/index.css";
   </style>
   <script type="module">
-    import '/js/index.ts'
+    import '/js/build/index.js'
   </script>
 
   <title>Commandtechno</title>
