+++ js/src/components/Activity/index.tsx
@@ -1,0 +1,3 @@
+export * from "./Activity";
+export * from "./ProgressBar";
+export * from "./types";
