+++ js/src/components/Post/index.tsx
@@ -1,0 +1,2 @@
+export * from "./Post";
+export * from "./types";
