+++ js/src/util/animate.ts
