+++ js/src/util/canScroll.ts
