+++ js/src/util/hide.ts
