+++ js/src/util/isRoot.ts
