--- tsconfig.json
@@ -1,6 +1,0 @@
-{
-  "compilerOptions": {
-    "jsx": "react",
-    "jsxFactory": "jsx"
-  }
-}
